chore: modify build process to use precompiled flatbuffers schema
This commit removes schema compilation, but instead uses a schema from
the releases section over at `rman-schema`. Submodule `rman-schema` is
no longer being used, however it will stay linked and updated as a way
to track which version of the schema is being used.
