+++ Cargo.toml
@@ -18,6 +18,10 @@ serde = { version = "1.0.152", features = ["derive"], optional = true }
 thiserror = "1.0.38"
 zstd = "0.12.1"
 
+[build-dependencies]
+flatbuffers = "22.10.26"
+zstd = "0.12.1"
+
 [dev-dependencies]
 httptest = "0.15.4"
 tokio = { version = "1.25.0", default_features = false, features = ["macros", "rt-multi-thread"] }
