+++ src/generated.rs
@@ -1,0 +1,6 @@
+#[allow(warnings)]
+#[allow(missing_debug_implementations, missing_copy_implementations)]
+#[allow(clippy::all, clippy::pedantic, clippy::nursery, clippy::unwrap_used)]
+mod schema_generated;
+
+pub use schema_generated::rman;
